+++ src/AutoRest.CSharp/LowLevel/AutoRest/LowLevelTarget.cs
@@ -13,7 +13,7 @@ namespace AutoRest.CSharp.AutoRest.Plugins
     {
         public static void Execute(GeneratedCodeWorkspace project, InputNamespace inputNamespace, SourceInputModel? sourceInputModel, bool cadlInput)
         {
-            var library = new DpgOutputLibraryBuilder(inputNamespace, sourceInputModel).Build(cadlInput);
+            var library = new DpgOutputLibrary(inputNamespace, cadlInput, sourceInputModel);
 
             //foreach (var enumType in library.Enums)
             //{
