+++ src/AutoRest.CSharp/LowLevel/Output/DpgOutputLibraryBuilder.cs
@@ -19,353 +19,353 @@ using Configuration = AutoRest.CSharp.Input.Configuration;
 
 namespace AutoRest.CSharp.Output.Models
 {
-    internal class DpgOutputLibraryBuilder
-    {
-        private const string MaxCountParameterName = "maxCount";
+    //internal class DpgOutputLibraryBuilder
+    //{
+    //    private const string MaxCountParameterName = "maxCount";
 
-        private readonly InputNamespace _rootNamespace;
-        private readonly SourceInputModel? _sourceInputModel;
-        private readonly string _defaultNamespace;
-        private readonly string _libraryName;
+    //    private readonly InputNamespace _rootNamespace;
+    //    private readonly SourceInputModel? _sourceInputModel;
+    //    private readonly string _defaultNamespace;
+    //    private readonly string _libraryName;
 
-        public DpgOutputLibraryBuilder(InputNamespace rootNamespace, SourceInputModel? sourceInputModel)
-        {
-            _rootNamespace = rootNamespace;
-            _sourceInputModel = sourceInputModel;
-            _defaultNamespace = Configuration.Namespace ?? rootNamespace.Name;
-            _libraryName = Configuration.LibraryName ?? rootNamespace.Name;
-        }
+    //    public DpgOutputLibraryBuilder(InputNamespace rootNamespace, SourceInputModel? sourceInputModel)
+    //    {
+    //        _rootNamespace = rootNamespace;
+    //        _sourceInputModel = sourceInputModel;
+    //        _defaultNamespace = Configuration.Namespace ?? rootNamespace.Name;
+    //        _libraryName = Configuration.LibraryName ?? rootNamespace.Name;
+    //    }
 
-        public DpgOutputLibrary Build(bool isCadlInput)
-        {
-            var inputClients = UpdateOperations();
+    //    public DpgOutputLibrary Build(bool isCadlInput)
+    //    {
+    //        var inputClients = UpdateOperations();
 
-            var clientInfosByName = inputClients
-                .Select(og => CreateClientInfo(og, _sourceInputModel, _rootNamespace.Name))
-                .ToDictionary(ci => ci.Name);
+    //        var clientInfosByName = inputClients
+    //            .Select(og => CreateClientInfo(og, _sourceInputModel, _rootNamespace.Name))
+    //            .ToDictionary(ci => ci.Name);
 
-            var topLevelClientInfos = SetHierarchy(clientInfosByName);
-            var clientOptions = CreateClientOptions(topLevelClientInfos);
+    //        var topLevelClientInfos = SetHierarchy(clientInfosByName);
+    //        var clientOptions = CreateClientOptions(topLevelClientInfos);
 
-            SetRequestsToClients(clientInfosByName.Values);
+    //        SetRequestsToClients(clientInfosByName.Values);
 
-            var enums = new Dictionary<InputEnumType, EnumType>(InputEnumType.IgnoreNullabilityComparer);
-            var models = new Dictionary<InputModelType, ObjectType>();
-            var clients = new List<LowLevelClient>();
+    //        var enums = new Dictionary<InputEnumType, EnumType>(InputEnumType.IgnoreNullabilityComparer);
+    //        var models = new Dictionary<InputModelType, ObjectType>();
+    //        var clients = new List<LowLevelClient>();
 
-            var library = new DpgOutputLibrary(enums, models, clients, clientOptions, isCadlInput);
+    //        var library = new DpgOutputLibrary(enums, models, clients, clientOptions, isCadlInput);
 
-            if (isCadlInput)
-            {
-                CreateEnums(enums, library.TypeFactory);
-                CreateModels(models, library.TypeFactory);
-            }
-            CreateClients(clients, topLevelClientInfos, library.TypeFactory, clientOptions);
+    //        if (isCadlInput)
+    //        {
+    //            CreateEnums(enums, library.TypeFactory);
+    //            CreateModels(models, library.TypeFactory);
+    //        }
+    //        CreateClients(clients, topLevelClientInfos, library.TypeFactory, clientOptions);
 
-            return library;
-        }
+    //        return library;
+    //    }
 
-        private void CreateEnums(IDictionary<InputEnumType, EnumType> dictionary, TypeFactory typeFactory)
-        {
-            foreach (var inputEnum in _rootNamespace.Enums)
-            {
-                if (inputEnum.Usage != InputModelTypeUsage.None)
-                {
-                    dictionary.Add(inputEnum, new EnumType(inputEnum, _defaultNamespace, "public", typeFactory, _sourceInputModel));
-                }
-            }
-        }
+    //    private void CreateEnums(IDictionary<InputEnumType, EnumType> dictionary, TypeFactory typeFactory)
+    //    {
+    //        foreach (var inputEnum in _rootNamespace.Enums)
+    //        {
+    //            if (inputEnum.Usage != InputModelTypeUsage.None)
+    //            {
+    //                dictionary.Add(inputEnum, new EnumType(inputEnum, _defaultNamespace, "public", typeFactory, _sourceInputModel));
+    //            }
+    //        }
+    //    }
 
-        private void CreateModels(IDictionary<InputModelType, ObjectType> models, TypeFactory typeFactory)
-        {
-            foreach (var model in _rootNamespace.Models)
-            {
-                if (model.Usage != InputModelTypeUsage.None)
-                {
-                    models.Add(model, new ModelTypeProvider(model, _defaultNamespace, _sourceInputModel, typeFactory));
-                }
-            }
-        }
+    //    private void CreateModels(IDictionary<InputModelType, ObjectType> models, TypeFactory typeFactory)
+    //    {
+    //        foreach (var model in _rootNamespace.Models)
+    //        {
+    //            if (model.Usage != InputModelTypeUsage.None)
+    //            {
+    //                models.Add(model, new ModelTypeProvider(model, _defaultNamespace, _sourceInputModel, typeFactory));
+    //            }
+    //        }
+    //    }
 
-        private IEnumerable<InputClient> UpdateOperations()
-        {
-            var defaultName = _rootNamespace.Name.ReplaceLast("Client", "");
-            // this map of old/new InputOperation is to update the lazy initialization of `Paging.NextLinkOperation`
-            var operationsMap = new Dictionary<InputOperation, InputOperation>();
-            foreach (var client in _rootNamespace.Clients)
-            {
-                var clientName = client.Name.IsNullOrEmpty() ? defaultName : client.Name;
-                yield return client with { Operations = client.Operations.Select(op => UpdateOperation(op, clientName, operationsMap)).ToList() };
-            }
-        }
+    //    private IEnumerable<InputClient> UpdateOperations()
+    //    {
+    //        var defaultName = _rootNamespace.Name.ReplaceLast("Client", "");
+    //        // this map of old/new InputOperation is to update the lazy initialization of `Paging.NextLinkOperation`
+    //        var operationsMap = new Dictionary<InputOperation, InputOperation>();
+    //        foreach (var client in _rootNamespace.Clients)
+    //        {
+    //            var clientName = client.Name.IsNullOrEmpty() ? defaultName : client.Name;
+    //            yield return client with { Operations = client.Operations.Select(op => UpdateOperation(op, clientName, operationsMap)).ToList() };
+    //        }
+    //    }
 
-        private static InputOperation UpdateOperation(InputOperation operation, string clientName, IDictionary<InputOperation, InputOperation> operationsMap)
-        {
-            InputOperation updatedOperation;
-            if (operation.Paging != null && !Configuration.DisablePaginationTopRenaming && !operation.Parameters.Any(p => p.Name.Equals(MaxCountParameterName, StringComparison.OrdinalIgnoreCase)))
-            {
-                updatedOperation = operation with
-                {
-                    Name = UpdateOperationName(operation, clientName),
-                    Parameters = UpdateOperationParameters(operation.Parameters),
-                    // to update the lazy initialization of `Paging.NextLinkOperation`
-                    Paging = operation.Paging with { NextLinkOperationRef = operation.Paging.NextLinkOperation != null ? () => operationsMap[operation.Paging.NextLinkOperation] : null }
-                };
-            }
-            else
-            {
-                updatedOperation = operation with { Name = UpdateOperationName(operation, clientName) };
-            }
-            operationsMap.Add(operation, updatedOperation);
+    //    private static InputOperation UpdateOperation(InputOperation operation, string clientName, IDictionary<InputOperation, InputOperation> operationsMap)
+    //    {
+    //        InputOperation updatedOperation;
+    //        if (operation.Paging != null && !Configuration.DisablePaginationTopRenaming && !operation.Parameters.Any(p => p.Name.Equals(MaxCountParameterName, StringComparison.OrdinalIgnoreCase)))
+    //        {
+    //            updatedOperation = operation with
+    //            {
+    //                Name = UpdateOperationName(operation, clientName),
+    //                Parameters = UpdateOperationParameters(operation.Parameters),
+    //                // to update the lazy initialization of `Paging.NextLinkOperation`
+    //                Paging = operation.Paging with { NextLinkOperationRef = operation.Paging.NextLinkOperation != null ? () => operationsMap[operation.Paging.NextLinkOperation] : null }
+    //            };
+    //        }
+    //        else
+    //        {
+    //            updatedOperation = operation with { Name = UpdateOperationName(operation, clientName) };
+    //        }
+    //        operationsMap.Add(operation, updatedOperation);
 
-            return updatedOperation;
+    //        return updatedOperation;
 
-        }
+    //    }
 
-        private static string UpdateOperationName(InputOperation operation, string clientName)
-            => operation.Name.RenameGetMethod(clientName).RenameListToGet(clientName);
+    //    private static string UpdateOperationName(InputOperation operation, string clientName)
+    //        => operation.Name.RenameGetMethod(clientName).RenameListToGet(clientName);
 
-        private static IReadOnlyList<InputParameter> UpdateOperationParameters(IReadOnlyList<InputParameter> operationParameters)
-        {
-            var parameters = new List<InputParameter>(operationParameters.Count);
-            foreach (var parameter in operationParameters)
-            {
-                if (parameter.Name.Equals("top", StringComparison.OrdinalIgnoreCase))
-                {
-                    parameters.Add(parameter with { Name = MaxCountParameterName });
-                }
-                else
-                {
-                    parameters.Add(parameter);
-                }
-            }
+    //    private static IReadOnlyList<InputParameter> UpdateOperationParameters(IReadOnlyList<InputParameter> operationParameters)
+    //    {
+    //        var parameters = new List<InputParameter>(operationParameters.Count);
+    //        foreach (var parameter in operationParameters)
+    //        {
+    //            if (parameter.Name.Equals("top", StringComparison.OrdinalIgnoreCase))
+    //            {
+    //                parameters.Add(parameter with { Name = MaxCountParameterName });
+    //            }
+    //            else
+    //            {
+    //                parameters.Add(parameter);
+    //            }
+    //        }
 
-            return parameters;
-        }
+    //        return parameters;
+    //    }
 
-        private ClientOptionsTypeProvider CreateClientOptions(IReadOnlyList<ClientInfo> topLevelClientInfos)
-        {
-            var clientName = topLevelClientInfos.Count == 1
-                ? topLevelClientInfos[0].Name
-                : topLevelClientInfos.SingleOrDefault(c => string.IsNullOrEmpty(c.OperationGroupKey))?.Name;
+    //    private ClientOptionsTypeProvider CreateClientOptions(IReadOnlyList<ClientInfo> topLevelClientInfos)
+    //    {
+    //        var clientName = topLevelClientInfos.Count == 1
+    //            ? topLevelClientInfos[0].Name
+    //            : topLevelClientInfos.SingleOrDefault(c => string.IsNullOrEmpty(c.OperationGroupKey))?.Name;
 
-            var clientOptionsName = $"{ClientBuilder.GetClientPrefix(clientName ?? _libraryName, _rootNamespace.Name)}ClientOptions";
-            var description = clientName != null
-                ? (FormattableString)$"Client options for {clientName}."
-                : $"Client options for {_libraryName} library clients.";
+    //        var clientOptionsName = $"{ClientBuilder.GetClientPrefix(clientName ?? _libraryName, _rootNamespace.Name)}ClientOptions";
+    //        var description = clientName != null
+    //            ? (FormattableString)$"Client options for {clientName}."
+    //            : $"Client options for {_libraryName} library clients.";
 
-            return new ClientOptionsTypeProvider(_sourceInputModel?.GetServiceVersionOverrides() ?? _rootNamespace.ApiVersions, clientOptionsName, _defaultNamespace, description, _sourceInputModel);
-        }
+    //        return new ClientOptionsTypeProvider(_sourceInputModel?.GetServiceVersionOverrides() ?? _rootNamespace.ApiVersions, clientOptionsName, _defaultNamespace, description, _sourceInputModel);
+    //    }
 
-        private static ClientInfo CreateClientInfo(InputClient ns, SourceInputModel? sourceInputModel, string rootNamespaceName)
-        {
-            var clientNamePrefix = ClientBuilder.GetClientPrefix(ns.Name, rootNamespaceName);
-            var clientNamespace = Configuration.Namespace ?? rootNamespaceName;
-            var clientDescription = ns.Description;
-            var operations = ns.Operations;
-            var clientParameters = RestClientBuilder.GetParametersFromOperations(operations).ToList();
-            var resourceParameters = clientParameters.Where(cp => cp.IsResourceParameter).ToHashSet();
-            var isSubClient = Configuration.SingleTopLevelClient && !string.IsNullOrEmpty(ns.Name) || resourceParameters.Any();
-            var clientName = isSubClient ? clientNamePrefix : clientNamePrefix + ClientBuilder.GetClientSuffix();
+    //    private static ClientInfo CreateClientInfo(InputClient ns, SourceInputModel? sourceInputModel, string rootNamespaceName)
+    //    {
+    //        var clientNamePrefix = ClientBuilder.GetClientPrefix(ns.Name, rootNamespaceName);
+    //        var clientNamespace = Configuration.Namespace ?? rootNamespaceName;
+    //        var clientDescription = ns.Description;
+    //        var operations = ns.Operations;
+    //        var clientParameters = RestClientBuilder.GetParametersFromOperations(operations).ToList();
+    //        var resourceParameters = clientParameters.Where(cp => cp.IsResourceParameter).ToHashSet();
+    //        var isSubClient = Configuration.SingleTopLevelClient && !string.IsNullOrEmpty(ns.Name) || resourceParameters.Any();
+    //        var clientName = isSubClient ? clientNamePrefix : clientNamePrefix + ClientBuilder.GetClientSuffix();
 
-            INamedTypeSymbol? existingType;
-            if (sourceInputModel == null || (existingType = sourceInputModel.FindForType(clientNamespace, clientName)) == null)
-            {
-                return new ClientInfo(ns.Name, clientName, clientNamespace, clientDescription, operations, clientParameters, resourceParameters);
-            }
+    //        INamedTypeSymbol? existingType;
+    //        if (sourceInputModel == null || (existingType = sourceInputModel.FindForType(clientNamespace, clientName)) == null)
+    //        {
+    //            return new ClientInfo(ns.Name, clientName, clientNamespace, clientDescription, operations, clientParameters, resourceParameters);
+    //        }
 
-            clientName = existingType.Name;
-            clientNamespace = existingType.ContainingNamespace.ToDisplayString();
-            return new ClientInfo(ns.Name, clientName, clientNamespace, clientDescription, existingType, operations, clientParameters, resourceParameters);
-        }
+    //        clientName = existingType.Name;
+    //        clientNamespace = existingType.ContainingNamespace.ToDisplayString();
+    //        return new ClientInfo(ns.Name, clientName, clientNamespace, clientDescription, existingType, operations, clientParameters, resourceParameters);
+    //    }
 
-        private IReadOnlyList<ClientInfo> SetHierarchy(IReadOnlyDictionary<string, ClientInfo> clientInfosByName)
-        {
-            if (_sourceInputModel != null)
-            {
-                foreach (var clientInfo in clientInfosByName.Values)
-                {
-                    AssignParents(clientInfo, clientInfosByName, _sourceInputModel);
-                }
-            }
+    //    private IReadOnlyList<ClientInfo> SetHierarchy(IReadOnlyDictionary<string, ClientInfo> clientInfosByName)
+    //    {
+    //        if (_sourceInputModel != null)
+    //        {
+    //            foreach (var clientInfo in clientInfosByName.Values)
+    //            {
+    //                AssignParents(clientInfo, clientInfosByName, _sourceInputModel);
+    //            }
+    //        }
 
-            var topLevelClients = clientInfosByName.Values.Where(c => c.Parent == null).ToList();
-            if (!Configuration.SingleTopLevelClient || topLevelClients.Count == 1)
-            {
-                return topLevelClients;
-            }
+    //        var topLevelClients = clientInfosByName.Values.Where(c => c.Parent == null).ToList();
+    //        if (!Configuration.SingleTopLevelClient || topLevelClients.Count == 1)
+    //        {
+    //            return topLevelClients;
+    //        }
 
-            var topLevelClientInfo = topLevelClients.FirstOrDefault(c => string.IsNullOrEmpty(c.OperationGroupKey));
-            if (topLevelClientInfo == null)
-            {
-                var clientName = ClientBuilder.GetClientPrefix(_libraryName, _rootNamespace.Name) + ClientBuilder.GetClientSuffix();
-                var clientNamespace = _defaultNamespace;
-                var endpointParameter = topLevelClients.SelectMany(c => c.ClientParameters).FirstOrDefault(p => p.IsEndpoint);
-                var clientParameters = endpointParameter != null ? new[] { endpointParameter } : Array.Empty<InputParameter>();
+    //        var topLevelClientInfo = topLevelClients.FirstOrDefault(c => string.IsNullOrEmpty(c.OperationGroupKey));
+    //        if (topLevelClientInfo == null)
+    //        {
+    //            var clientName = ClientBuilder.GetClientPrefix(_libraryName, _rootNamespace.Name) + ClientBuilder.GetClientSuffix();
+    //            var clientNamespace = _defaultNamespace;
+    //            var endpointParameter = topLevelClients.SelectMany(c => c.ClientParameters).FirstOrDefault(p => p.IsEndpoint);
+    //            var clientParameters = endpointParameter != null ? new[] { endpointParameter } : Array.Empty<InputParameter>();
 
-                topLevelClientInfo = new ClientInfo(clientName, clientNamespace, clientParameters);
-            }
+    //            topLevelClientInfo = new ClientInfo(clientName, clientNamespace, clientParameters);
+    //        }
 
-            foreach (var clientInfo in topLevelClients)
-            {
-                if (clientInfo != topLevelClientInfo)
-                {
-                    clientInfo.Parent = topLevelClientInfo;
-                    topLevelClientInfo.Children.Add(clientInfo);
-                }
-            }
+    //        foreach (var clientInfo in topLevelClients)
+    //        {
+    //            if (clientInfo != topLevelClientInfo)
+    //            {
+    //                clientInfo.Parent = topLevelClientInfo;
+    //                topLevelClientInfo.Children.Add(clientInfo);
+    //            }
+    //        }
 
-            return new[] { topLevelClientInfo };
-        }
+    //        return new[] { topLevelClientInfo };
+    //    }
 
-        private static void AssignParents(in ClientInfo clientInfo, IReadOnlyDictionary<string, ClientInfo> clientInfosByName, SourceInputModel sourceInputModel)
-        {
-            var child = clientInfo;
-            while (child.Parent == null && child.ExistingType != null && sourceInputModel.TryGetClientSourceInput(child.ExistingType, out var clientSourceInput) && clientSourceInput.ParentClientType != null)
-            {
-                var parent = clientInfosByName[clientSourceInput.ParentClientType.Name];
-                if (clientInfo == parent)
-                {
-                    throw new InvalidOperationException("loop");
-                }
+    //    private static void AssignParents(in ClientInfo clientInfo, IReadOnlyDictionary<string, ClientInfo> clientInfosByName, SourceInputModel sourceInputModel)
+    //    {
+    //        var child = clientInfo;
+    //        while (child.Parent == null && child.ExistingType != null && sourceInputModel.TryGetClientSourceInput(child.ExistingType, out var clientSourceInput) && clientSourceInput.ParentClientType != null)
+    //        {
+    //            var parent = clientInfosByName[clientSourceInput.ParentClientType.Name];
+    //            if (clientInfo == parent)
+    //            {
+    //                throw new InvalidOperationException("loop");
+    //            }
 
-                child.Parent = parent;
-                parent.Children.Add(child);
-                child = parent;
-            }
-        }
+    //            child.Parent = parent;
+    //            parent.Children.Add(child);
+    //            child = parent;
+    //        }
+    //    }
 
-        private static void SetRequestsToClients(IEnumerable<ClientInfo> clientInfos)
-        {
-            foreach (var clientInfo in clientInfos)
-            {
-                foreach (var operation in clientInfo.Operations)
-                {
-                    SetRequestToClient(clientInfo, operation);
-                }
-            }
-        }
+    //    private static void SetRequestsToClients(IEnumerable<ClientInfo> clientInfos)
+    //    {
+    //        foreach (var clientInfo in clientInfos)
+    //        {
+    //            foreach (var operation in clientInfo.Operations)
+    //            {
+    //                SetRequestToClient(clientInfo, operation);
+    //            }
+    //        }
+    //    }
 
-        private static void SetRequestToClient(ClientInfo clientInfo, InputOperation operation)
-        {
-            switch (clientInfo.ResourceParameters.Count)
-            {
-                case 1:
-                    var requestParameter = clientInfo.ResourceParameters.Single();
-                    if (operation.Parameters.Contains(requestParameter))
-                    {
-                        break;
-                    }
+    //    private static void SetRequestToClient(ClientInfo clientInfo, InputOperation operation)
+    //    {
+    //        switch (clientInfo.ResourceParameters.Count)
+    //        {
+    //            case 1:
+    //                var requestParameter = clientInfo.ResourceParameters.Single();
+    //                if (operation.Parameters.Contains(requestParameter))
+    //                {
+    //                    break;
+    //                }
 
-                    while (clientInfo.Parent != null && clientInfo.ResourceParameters.Count != 0)
-                    {
-                        clientInfo = clientInfo.Parent;
-                    }
-                    break;
-                case > 1:
-                    var requestParameters = operation.Parameters.ToHashSet();
-                    while (clientInfo.Parent != null && !clientInfo.ResourceParameters.IsSubsetOf(requestParameters))
-                    {
-                        clientInfo = clientInfo.Parent;
-                    }
+    //                while (clientInfo.Parent != null && clientInfo.ResourceParameters.Count != 0)
+    //                {
+    //                    clientInfo = clientInfo.Parent;
+    //                }
+    //                break;
+    //            case > 1:
+    //                var requestParameters = operation.Parameters.ToHashSet();
+    //                while (clientInfo.Parent != null && !clientInfo.ResourceParameters.IsSubsetOf(requestParameters))
+    //                {
+    //                    clientInfo = clientInfo.Parent;
+    //                }
 
-                    break;
-            }
+    //                break;
+    //        }
 
-            clientInfo.Requests.Add(operation);
-        }
+    //        clientInfo.Requests.Add(operation);
+    //    }
 
-        private void CreateClients(List<LowLevelClient> allClients, IEnumerable<ClientInfo> topLevelClientInfos, TypeFactory typeFactory, ClientOptionsTypeProvider clientOptions)
-        {
-            var topLevelClients = CreateClients(topLevelClientInfos, typeFactory, clientOptions, null);
+    //    private void CreateClients(List<LowLevelClient> allClients, IEnumerable<ClientInfo> topLevelClientInfos, TypeFactory typeFactory, ClientOptionsTypeProvider clientOptions)
+    //    {
+    //        var topLevelClients = CreateClients(topLevelClientInfos, typeFactory, clientOptions, null);
 
-            // Simple implementation of breadth first traversal
-            allClients.AddRange(topLevelClients);
-            for (int i = 0; i < allClients.Count; i++)
-            {
-                allClients.AddRange(allClients[i].SubClients);
-            }
-        }
+    //        // Simple implementation of breadth first traversal
+    //        allClients.AddRange(topLevelClients);
+    //        for (int i = 0; i < allClients.Count; i++)
+    //        {
+    //            allClients.AddRange(allClients[i].SubClients);
+    //        }
+    //    }
 
-        private IEnumerable<LowLevelClient> CreateClients(IEnumerable<ClientInfo> clientInfos, TypeFactory typeFactory, ClientOptionsTypeProvider clientOptions, LowLevelClient? parentClient)
-        {
-            foreach (var clientInfo in clientInfos)
-            {
-                var description = string.IsNullOrWhiteSpace(clientInfo.Description)
-                    ? $"The {ClientBuilder.GetClientPrefix(clientInfo.Name, _rootNamespace.Name)} {(parentClient == null ? "service client" : "sub-client")}."
-                    : BuilderHelpers.EscapeXmlDescription(clientInfo.Description);
+    //    private IEnumerable<LowLevelClient> CreateClients(IEnumerable<ClientInfo> clientInfos, TypeFactory typeFactory, ClientOptionsTypeProvider clientOptions, LowLevelClient? parentClient)
+    //    {
+    //        foreach (var clientInfo in clientInfos)
+    //        {
+    //            var description = string.IsNullOrWhiteSpace(clientInfo.Description)
+    //                ? $"The {ClientBuilder.GetClientPrefix(clientInfo.Name, _rootNamespace.Name)} {(parentClient == null ? "service client" : "sub-client")}."
+    //                : BuilderHelpers.EscapeXmlDescription(clientInfo.Description);
 
-                var subClients = new List<LowLevelClient>();
+    //            var subClients = new List<LowLevelClient>();
 
-                var client = new LowLevelClient(
-                    clientInfo.Name,
-                    clientInfo.Namespace,
-                    description,
-                    _libraryName,
-                    parentClient,
-                    clientInfo.Requests,
-                    clientInfo.ClientParameters,
-                    _rootNamespace.Auth,
-                    _sourceInputModel,
-                    clientOptions,
-                    typeFactory)
-                {
-                    SubClients = subClients
-                };
+    //            var client = new LowLevelClient(
+    //                clientInfo.Name,
+    //                clientInfo.Namespace,
+    //                description,
+    //                _libraryName,
+    //                parentClient,
+    //                clientInfo.Requests,
+    //                clientInfo.ClientParameters,
+    //                _rootNamespace.Auth,
+    //                _sourceInputModel,
+    //                clientOptions,
+    //                typeFactory)
+    //            {
+    //                SubClients = subClients
+    //            };
 
-                subClients.AddRange(CreateClients(clientInfo.Children, typeFactory, clientOptions, client));
+    //            subClients.AddRange(CreateClients(clientInfo.Children, typeFactory, clientOptions, client));
 
-                yield return client;
-            }
-        }
+    //            yield return client;
+    //        }
+    //    }
 
-        private class ClientInfo
-        {
-            public string OperationGroupKey { get; }
-            public string Name { get; }
-            public string Namespace { get; }
-            public string Description { get; }
-            public INamedTypeSymbol? ExistingType { get; }
-            public IReadOnlyList<InputOperation> Operations { get; }
-            public IReadOnlyList<InputParameter> ClientParameters { get; }
-            public ISet<InputParameter> ResourceParameters { get; }
+    //    private class ClientInfo
+    //    {
+    //        public string OperationGroupKey { get; }
+    //        public string Name { get; }
+    //        public string Namespace { get; }
+    //        public string Description { get; }
+    //        public INamedTypeSymbol? ExistingType { get; }
+    //        public IReadOnlyList<InputOperation> Operations { get; }
+    //        public IReadOnlyList<InputParameter> ClientParameters { get; }
+    //        public ISet<InputParameter> ResourceParameters { get; }
 
-            public ClientInfo? Parent { get; set; }
-            public IList<ClientInfo> Children { get; }
-            public IList<InputOperation> Requests { get; }
+    //        public ClientInfo? Parent { get; set; }
+    //        public IList<ClientInfo> Children { get; }
+    //        public IList<InputOperation> Requests { get; }
 
-            public ClientInfo(string operationGroupKey, string clientName, string clientNamespace, string clientDescription, IReadOnlyList<InputOperation> operations, IReadOnlyList<InputParameter> clientParameters, ISet<InputParameter> resourceParameters)
-                : this(operationGroupKey, clientName, clientNamespace, clientDescription, null, operations, clientParameters, resourceParameters)
-            {
-            }
+    //        public ClientInfo(string operationGroupKey, string clientName, string clientNamespace, string clientDescription, IReadOnlyList<InputOperation> operations, IReadOnlyList<InputParameter> clientParameters, ISet<InputParameter> resourceParameters)
+    //            : this(operationGroupKey, clientName, clientNamespace, clientDescription, null, operations, clientParameters, resourceParameters)
+    //        {
+    //        }
 
-            public ClientInfo(string operationGroupKey, string clientName, string clientNamespace, string clientDescription, INamedTypeSymbol? existingType, IReadOnlyList<InputOperation> operations, IReadOnlyList<InputParameter> clientParameters, ISet<InputParameter> resourceParameters)
-            {
-                OperationGroupKey = operationGroupKey;
-                Name = clientName;
-                Namespace = clientNamespace;
-                Description = clientDescription;
-                ExistingType = existingType;
-                Operations = operations;
-                ClientParameters = clientParameters;
-                ResourceParameters = resourceParameters;
-                Children = new List<ClientInfo>();
-                Requests = new List<InputOperation>();
-            }
+    //        public ClientInfo(string operationGroupKey, string clientName, string clientNamespace, string clientDescription, INamedTypeSymbol? existingType, IReadOnlyList<InputOperation> operations, IReadOnlyList<InputParameter> clientParameters, ISet<InputParameter> resourceParameters)
+    //        {
+    //            OperationGroupKey = operationGroupKey;
+    //            Name = clientName;
+    //            Namespace = clientNamespace;
+    //            Description = clientDescription;
+    //            ExistingType = existingType;
+    //            Operations = operations;
+    //            ClientParameters = clientParameters;
+    //            ResourceParameters = resourceParameters;
+    //            Children = new List<ClientInfo>();
+    //            Requests = new List<InputOperation>();
+    //        }
 
-            public ClientInfo(string clientName, string clientNamespace, IReadOnlyList<InputParameter> clientParameters)
-            {
-                OperationGroupKey = string.Empty;
-                Name = clientName;
-                Namespace = clientNamespace;
-                Description = string.Empty;
-                ExistingType = null;
-                Operations = Array.Empty<InputOperation>();
-                ClientParameters = clientParameters;
-                ResourceParameters = new HashSet<InputParameter>();
-                Children = new List<ClientInfo>();
-                Requests = new List<InputOperation>();
-            }
-        }
-    }
+    //        public ClientInfo(string clientName, string clientNamespace, IReadOnlyList<InputParameter> clientParameters)
+    //        {
+    //            OperationGroupKey = string.Empty;
+    //            Name = clientName;
+    //            Namespace = clientNamespace;
+    //            Description = string.Empty;
+    //            ExistingType = null;
+    //            Operations = Array.Empty<InputOperation>();
+    //            ClientParameters = clientParameters;
+    //            ResourceParameters = new HashSet<InputParameter>();
+    //            Children = new List<ClientInfo>();
+    //            Requests = new List<InputOperation>();
+    //        }
+    //    }
+    //}
 }
